+++ src/fixed_width_index_connection.h
@@ -14,7 +14,7 @@ public:
       std::vector<int> col_ends,
       bool trim_ws,
       const size_t skip,
-      const char comment,
+      const char* comment,
       const size_t n_max,
       const bool progress,
       const size_t chunk_size);